feat: 이미지 관련 예외 처리

+++ src/main/java/com/leets/team2/xclone/domain/post/service/PostService.java
@@ -4,6 +4,7 @@ import com.leets.team2.xclone.domain.member.entities.Member;
 import com.leets.team2.xclone.domain.post.dto.PostDTO;
 import com.leets.team2.xclone.domain.post.entity.Post;
 import com.leets.team2.xclone.domain.post.repository.PostRepository;
+import com.leets.team2.xclone.exception.InvalidFileException;
 import com.leets.team2.xclone.exception.PostNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -24,6 +25,10 @@ public class PostService {
         String imageUrl=null;//이미지 url 초기화
 
         if(imageFile!=null&&!imageFile.isEmpty()){
+            if(imageFile.getContentType()==null||!imageFile.getContentType().equals("image/png")){
+                throw new InvalidFileException();
+            }
+
             String imageName= UUID.randomUUID().toString()+"_"+imageFile.getOriginalFilename();//이미지 파일의 이름 생성
             Path imagePath= Paths.get("uploads/"+imageName);//저장 경로 설정
             Files.createDirectories(imagePath.getParent());//폴더 없으면 새로 생성
@@ -47,6 +52,9 @@ public class PostService {
         post.setContent(postDTO.getContent());
 
         if(imageFile!=null&&!imageFile.isEmpty()){//새로운 이미지가 있을 경우 원래 있던 이미지 삭제하고 업데이트한다.
+            if (imageFile.getContentType()==null||!imageFile.getContentType().equals("image/png")) {
+                throw new InvalidFileException();
+            }
             if(post.getImageUrl()!=null){
                 Path existingImagePath=Paths.get(post.getImageUrl());
                 Files.deleteIfExists(existingImagePath);//기존 파일 삭제(존재한다면)
